Add parallelism to the longitudinal pipeline (long_fastsurfer.sh)
This parallelizes the surface pipelines of all steps and timepoints wherever possible.
The segmentation pipeline is only parallelized insofar as it runs in parallel to the surfaces where possible, but only one segmentation at a time to avoid memory limits of gpus (one gpu is still enough).
This is not optimal w.r.t. GPU utilization, because it is not really clear for "outside processes" when the GPU is no longer being used (which is at the end of the Longitudinal Segmentation